Delete unnecessary updater
